requirements.txt
numpy
pandas
xgboost
scikit-learn
joblib
fastapi
uvicorn
requests
python-multipart
python-dotenv
twilio
